backend/requirements.txt
asgiref==3.5.2
bleach==5.0.1
Django==4.1
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
webencodings==0.5.1
